chore: Add coverage report to CI
