+++ crates/redis_config/tests/async_source/mod.rs
@@ -1,0 +1,1 @@
+mod test_async_source;
