+++ crates/redis_config/tests/main.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod async_source;
